+++ src/trio_util/__init__.py
@@ -1,6 +1,5 @@
 from ._version import __version__
 from ._async_bool import AsyncBool
-from ._async_dictionary import AsyncDictionary
 from ._async_itertools import azip, azip_longest
 from ._async_value import AsyncValue, compose_values
 from ._awaitables import wait_all, wait_any